+++ Source/Core/VideoBackends/D3D12/D3D12Renderer.h
@@ -5,6 +5,7 @@
 #pragma once
 
 #include <d3d12.h>
+#include "VideoBackends/D3D12/DescriptorAllocator.h"
 #include "VideoBackends/D3D12/DescriptorHeapManager.h"
 #include "VideoCommon/RenderBase.h"
