+++ Source/Core/VideoBackends/D3DCommon/D3DCommon.h
