+++ Source/Core/VideoBackends/D3DCommon/D3DCommon.vcxproj
@@ -16,12 +16,12 @@
   </ImportGroup>
   <PropertyGroup Label="UserMacros" />
   <ItemGroup>
-    <ClCompile Include="Common.cpp" />
+    <ClCompile Include="D3DCommon.cpp" />
     <ClCompile Include="Shader.cpp" />
     <ClCompile Include="SwapChain.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Common.h" />
+    <ClInclude Include="D3DCommon.h" />
     <ClInclude Include="Shader.h" />
     <ClInclude Include="SwapChain.h" />
   </ItemGroup>
