+++ Source/Core/VideoBackends/D3DCommon/SwapChain.h
@@ -9,7 +9,7 @@
 
 #include "Common/CommonTypes.h"
 #include "Common/WindowSystemInfo.h"
-#include "VideoBackends/D3DCommon/Common.h"
+#include "VideoBackends/D3DCommon/D3DCommon.h"
 #include "VideoCommon/TextureConfig.h"
 
 namespace D3DCommon
